[#86460670] Added suggested changes and a few tests

+++ lib/version_parser.rb
@@ -1,6 +1,10 @@
 
 class VersionParser
   def get_version_from_file filename
+
+    raise IOError,"The log \"#{filename}\" is not a MainOutput log file." unless filename.include? 'MainOutputLog_'
+    raise IOError,"The log \"#{filename}\" does not exist." unless File.exist? filename
+
     file = File.open(filename)
     file.each_line do |line|
       if line.include?('Version')
